requirements.txt
pandas
scikit-learn
transformers
nltk
torch
flask
joblib
streamlit
langchain_groq
python-dotenv